+++ resource/resource.py
@@ -1,0 +1,15 @@
+import json
+
+base = 35
+
+def get_activation_token_by_login(
+        login,
+        response
+):
+    token = None
+    for item in response.json()['items']:
+        user_data = json.loads(item['Content']['Body'])
+        user_login = user_data['Login']
+        if user_login == login:
+            token = user_data['ConfirmationLinkUrl'].split('/')[-1]
+            return token
